Refactor: SeatService - seatReception

+++ src/main/java/ticketaka/mtvs3_final_backend/seat/command/domain/repository/MemberSeatRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 @Repository
 public interface MemberSeatRepository extends JpaRepository<MemberSeat, Long> {
 
-    Long countByConcertIdAndSeatIdAndMemberSeatStatus(Long concertId, Long seatId, MemberSeatStatus status);
+    int countByConcertIdAndSeatIdAndMemberSeatStatus(Long concertId, Long seatId, MemberSeatStatus status);
 
     int countByMemberIdAndConcertId(Long memberId, Long concertId);
